@@ -7,8 +7,8 @@ apply plugin: 'signing'
 
 group = "ch.softappeal.yass"
 
-sourceCompatibility = '1.7'
-targetCompatibility = '1.7'
+sourceCompatibility = '1.8'
+targetCompatibility = '1.8'
 
 repositories {
   mavenCentral()
