@@ -5,7 +5,6 @@ import ch.softappeal.yass.serialize.Reflector;
 import java.lang.reflect.Field;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -21,11 +20,7 @@ public final class SimpleFastSerializer extends AbstractFastSerializer {
     int fieldId = FieldHandler.FIRST_ID;
     for (Class<?> t = type; (t != null) && (t != Throwable.class); t = t.getSuperclass()) {
       final List<Field> fields = ownFields(t);
-      Collections.sort(fields, new Comparator<Field>() {
-        @Override public int compare(final Field field1, final Field field2) {
-          return field1.getName().compareTo(field2.getName());
-        }
-      });
+      Collections.sort(fields, (field1, field2) -> field1.getName().compareTo(field2.getName()));
       for (final Field field : fields) {
         id2field.put(fieldId++, field);
       }
