@@ -26,14 +26,12 @@ public final class SocketExecutor {
 
   void execute(final Socket socket, final SocketListener listener) {
     try {
-      executor.execute(new Runnable() {
-        @Override public void run() {
-          try {
-            listener.accept(socket, executor);
-          } catch (final Exception e) {
-            SocketTransport.close(socket, e);
-            throw Exceptions.wrap(e);
-          }
+      executor.execute(() -> {
+        try {
+          listener.accept(socket, executor);
+        } catch (final Exception e) {
+          SocketTransport.close(socket, e);
+          throw Exceptions.wrap(e);
         }
       });
     } catch (final Exception e) {
