@@ -35,10 +35,6 @@ public final class TaggedMethodMapper implements MethodMapper {
     return id2mapping.get(method.getAnnotation(Tag.class).value());
   }
 
-  public static final Factory FACTORY = new Factory() {
-    @Override public MethodMapper create(final Class<?> contract) {
-      return new TaggedMethodMapper(contract);
-    }
-  };
+  public static final Factory FACTORY = TaggedMethodMapper::new;
 
 }
