@@ -22,18 +22,16 @@ public final class Exceptions {
    * <p/>
    * Note: For productive code, this handler should be replaced with one that uses your logging framework!
    */
-  public static final UncaughtExceptionHandler STD_ERR = new UncaughtExceptionHandler() {
-    @Override public void uncaughtException(final Thread thread, final Throwable throwable) {
-      System.err.println(
-        "### " + new Date() + " - " + ((thread == null) ? "<null>" : thread.getName()) + " - " + Exceptions.class.getName() + ':'
-      );
-      if (throwable == null) {
-        System.err.println("throwable is null");
-      } else {
-        throwable.printStackTrace();
-        if (!(throwable instanceof Exception)) {
-          System.exit(1);
-        }
+  public static final UncaughtExceptionHandler STD_ERR = (thread, throwable) -> {
+    System.err.println(
+      "### " + new Date() + " - " + ((thread == null) ? "<null>" : thread.getName()) + " - " + Exceptions.class.getName() + ':'
+    );
+    if (throwable == null) {
+      System.err.println("throwable is null");
+    } else {
+      throwable.printStackTrace();
+      if (!(throwable instanceof Exception)) {
+        System.exit(1);
       }
     }
   };
@@ -43,11 +41,9 @@ public final class Exceptions {
    * <p/>
    * Note: For productive code, this handler should be replaced with one that uses your logging framework!
    */
-  public static final UncaughtExceptionHandler TERMINATE = new UncaughtExceptionHandler() {
-    @Override public void uncaughtException(final Thread thread, final Throwable throwable) {
-      STD_ERR.uncaughtException(thread, throwable);
-      System.exit(1);
-    }
+  public static final UncaughtExceptionHandler TERMINATE = (thread, throwable) -> {
+    STD_ERR.uncaughtException(thread, throwable);
+    System.exit(1);
   };
 
 }
